requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
curl_cffi==0.13.0
fastapi==0.135.3
frozendict==2.4.7
greenlet==3.3.2
h11==0.16.0
idna==3.11
multitasking==0.0.12
numpy==2.4.4
pandas==3.0.2
peewee==4.0.4
platformdirs==4.9.4
protobuf==7.34.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.43.0
websockets==16.0
yfinance==1.2.0
fredapi==0.5.1
PyJWT[crypto]==2.10.1
httpx>=0.27
slowapi==0.1.9
